added logging function file so someone can work on it concurrently, also started to make fixed point math a thing and verified there's no overflow in the ones that have been converted

+++ PWM.hpp
@@ -5,7 +5,7 @@
 
 
 
-void setDutyCycle(uint8_t Channel, uint16_t duty_cycle) { //this function exists because analogWrite doesn't use OCR0x register. Or at least it doesn't look like it
+void setDutyCycle(uint8_t Channel, uint8_t duty_cycle) { //this function exists because analogWrite doesn't use OCR0x register. Or at least it doesn't look like it
 
 	if (Channel == 5) {
